requirements.txt
streamlit
google-generativeai
python-dotenv
openai-whisper
ffmpeg
